organized pages folder into match scouting and settings
