+++ website/src/pages/Settings/SettingsPage.jsx
@@ -1,12 +1,12 @@
 import React, { useEffect, useState } from "react";
 import { toast } from "react-toastify";
 
-import ProceedBackButton from "../components/ProceedBackButton";
-import SettingsMatchDataScanner from "../components/SettingsComponents/SettingsMatchDataScanner";
-import SettingsButton from "../components/SettingsComponents/SettingsButton";
-import SettingsViewMatchData from "../components/SettingsComponents/SettingsViewMatchData";
-import SettingsUpdateButton from "../components/SettingsComponents/SettingsUpdateButton";
-import { useNavigateWithBase } from "../utils/useNavigateWithBase";
+import ProceedBackButton from "../../components/ProceedBackButton";
+import SettingsMatchDataScanner from "../../components/SettingsComponents/SettingsMatchDataScanner";
+import SettingsButton from "../../components/SettingsComponents/SettingsButton";
+import SettingsViewMatchData from "../../components/SettingsComponents/SettingsViewMatchData";
+import SettingsUpdateButton from "../../components/SettingsComponents/SettingsUpdateButton";
+import { useNavigateWithBase } from "../../utils/useNavigateWithBase";
 
 /**
  * A page for the user to access settings such as clearing match data, viewing match data, and getting match data.
